backend/requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn>=0.21.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
upstash-vector>=0.1.0,<1.0.0
groq>=0.1.0,<1.0.0
pydantic>=1.10.0,<2.0.0
python-multipart>=0.0.5,<1.0.0
